+++ dist_app/sn/FocusMng.d.ts
@@ -1,8 +1,8 @@
+import { SysBase } from './SysBase';
 import { Container } from 'pixi.js';
 export declare class FocusMng {
     #private;
-    private readonly cvs;
-    constructor(cvs: HTMLCanvasElement);
+    constructor(cvs: HTMLCanvasElement, sys: SysBase);
     destroy(): void;
     add(cmp: Container | HTMLElement, on: () => boolean, off: () => void): void;
     remove(cmp: Container | HTMLElement): void;